Add a new asynchronous executor based on Trollius
The executor runs the asyncio event loop in a dedicated thread. It is possible
to specify an asyncio event loop to avoid this thread. For example, it is
possible to use the greenio module to create an event loop running in greenlet,
and so run the asyncio event loop in the main thread. The callback and
incoming.reply() are executed in the same thread than the event loop.

incoming.reply() can be called in a pool of thread, but this feature is
currently disabled (until I fix the stop issue).

The changeset implements the asyncio blueprint:
https://wiki.openstack.org/wiki/Oslo/blueprints/asyncio

Trollius is a port on Python 2 of the new asyncio module of Python 3.4 which is
also known as the Tulip project (written for Python 3.3 and later).
Documentation of the asyncio module:
http://docs.python.org/dev/library/asyncio.html

The patchset adds a new dependency, the "trollius" module. Trollius is already
part of the global requirements file:
https://review.openstack.org/#/c/70983/

Currently, my test program hangs at exit because the threading calling
listener.poll() is blocked.  I created the following issue to add a timeout
parameter to Listener.poll(), so the thread will be able to exit:
https://review.openstack.org/#/c/71003/

Change-Id: I3815673aad35c560d7e7df2fa2a6601a1bbfb773
Implements: blueprint asyncio

Conflicts:
	setup.cfg