Add transport reconnection retries
When a rpc client try to make a RPC call and the server is unreachable
The rpc call hang until the server come back.

In most case this is the desired behavior.

But sometimes, we can prefer that the library raise an exception after a
certain number of retries.

This patch allow this.

Closes bug #1282639

Change-Id: I32086d0abf141c368343bf225d4b021da496c020